+++ src/Hacks/TPSBypass.cpp
@@ -88,11 +88,11 @@ void updateTPSPatches(bool tpsEnabled)
     if (tpsEnabled)
     {
         #ifdef GEODE_IS_WINDOWS
-        //auto array = geode::toByteArray<float>(1.0f / tps);
+        auto array = geode::toByteArray<float>(1.0f / tps);
 
         //DWORD old_prot;
         //VirtualProtect(reinterpret_cast<void*>(geode::base::get() + 0x5ec6d0), array.size(), PAGE_EXECUTE_READWRITE, &old_prot);
-        //patches.push_back(createPatchSafe(reinterpret_cast<void*>(geode::base::get() + 0x5ec6d0), array));
+        patches.push_back(createPatchSafe(reinterpret_cast<void*>(geode::base::get() + 0x5ec6d0), array));
         //VirtualProtect(reinterpret_cast<void*>(geode::base::get() + 0x5ec6d0), array.size(), old_prot, &old_prot);
         #endif